Create Secret number
This code saves the nanoseconds since some fixed time to the pseudoRandomNumber variable. Now we can assume that this is a random number. So you can generate a secret number by iterating the pseudoRandomNumber in reverse order and adding unique digits in the secret number. If the pseudoRandomNumber lacks the required amount of unique digits, call System.nanoTime() again and try to generate the secret number again until you get a satisfactory number.